+++ index.ts
@@ -1,5 +1,5 @@
 import expectMessage from "./expectMessage/expectMessage";
-import { readCode } from "./readCode/readCode";
+export * from "./readCode/readCode";
 export * from "./astNodeParser/astNodeParser";
 
-export { expectMessage, readCode };
+export { expectMessage };
